Refactor: Split currency into separate module
